requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pytest>=7.0.0
pyyaml>=6.0
mlflow>=2.0.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
